Modernized and renamed brewing stand property
